+++ src/main/kotlin/cloud/glitchdev/rfu/gui/components/elementa/MultilineHudText.kt
@@ -1,5 +1,6 @@
 package cloud.glitchdev.rfu.gui.components.elementa
 
+import gg.essential.elementa.UIComponent
 import gg.essential.elementa.components.UIContainer
 import gg.essential.elementa.components.UIText
 import gg.essential.elementa.constraints.ChildBasedMaxSizeConstraint
@@ -16,7 +17,7 @@ import java.awt.Color
 class MultilineHudText(private var scale: Float) : UIContainer() {
     data class HudLine(val text: String, val color: Color = Color.WHITE)
 
-    private var lines: List<UIText> = emptyList()
+    private var lines: MutableList<UIComponent> = mutableListOf()
     private var currentText: String = ""
 
     init {
@@ -29,7 +30,8 @@ class MultilineHudText(private var scale: Float) : UIContainer() {
     fun setLines(hudLines: List<HudLine>) {
         currentText = hudLines.joinToString("\n") { it.text }
         lines.forEach { removeChild(it) }
-        lines = hudLines.map { line ->
+        lines.clear()
+        lines.addAll(hudLines.map { line ->
             UIText(line.text).constrain {
                 x = 0.pixels()
                 y = SiblingConstraint()
@@ -37,7 +39,7 @@ class MultilineHudText(private var scale: Float) : UIContainer() {
                 height = TextAspectConstraint()
                 this.color = line.color.toConstraint()
             } childOf this
-        }
+        })
     }
 
     fun setText(value: String) {
@@ -45,6 +47,38 @@ class MultilineHudText(private var scale: Float) : UIContainer() {
             setLines(emptyList())
         } else {
             setLines(value.split("\n").map { HudLine(it) })
+        }
+    }
+
+    fun clearLines() {
+        lines.forEach { removeChild(it) }
+        lines.clear()
+        currentText = ""
+    }
+
+    fun addLine(text: String, index: Int? = null) {
+        val uiText = UIText(text).constrain {
+            x = 0.pixels()
+            y = SiblingConstraint()
+            width = ScaledTextConstraint(scale)
+            height = TextAspectConstraint()
+        } childOf this
+        if (index != null) {
+            lines.add(index, uiText)
+        } else {
+            lines.add(uiText)
+        }
+    }
+
+    fun addLine(line: UIComponent, index: Int? = null) {
+        line.constrain {
+            x = 0.pixels()
+            y = SiblingConstraint()
+        } childOf this
+        if (index != null) {
+            lines.add(index, line)
+        } else {
+            lines.add(line)
         }
     }